linux-firmware: Create necessary symlinks for BT
When initializing bluetooth, it is necessary for
the firmware binary to be pointed to by a specific
symlink, so hciattach can work properly.

This was previously setup by our imx-firmware recipe,
but since the firmwares were removed by upstream, we
are now able to get it from the linux-firmware package.

Signed-off-by: Alexandru Costache <alexandru@balena.io>