resin-image: Install device-tree in boot partition
When performing an altboot test from kernel 4.9
to kernel 4.14.98 we notice altboot fails even
if kernel is loaded from rootfs. The reason
is the new 4.14.98 dtb is not compatible
with the old kernel, and causes a halt.

Let's fix this for further releases now
by loading the dtb from the rootfs too.

Signed-off-by: Alexandru Costache <alexandru@balena.io>