balena-image-initramfs: Remove unused modules
Because the device types in this repo are not a target
for migration. Also we have a dedicated debug uart on this
board so we can also remove the recovery module.
We need to remove these modules because we are running
out of space in the rootfs and this makes HUP error.

Changelog-entry: balena-image-initramfs: Remove unused modules
Signed-off-by: Florin Sarbu <florin@balena.io>